dados obrigatorios em login e cadastro

+++ src/app/login-page/cadastro-form/cadastro-form.component.ts
@@ -42,41 +42,48 @@ export class CadastroFormComponent {
 
   onSubmit(value: any) {
     event.preventDefault();
-    this.spinner.show();
-    this.user = new Usuario(
-      value.nome,
-      value.username,
-      value.email,
-      value.password,
-      0,
-      1,
-      'default.png',
-      true
-    );
-    this.httpClient
-      .post("http://monica:64803/api/Usuario", this.user, {
-        observe: "response"
-      })
-      .pipe(map(res => res as any))
-      .subscribe(
-        res => {
-          localStorage.setItem("userId", JSON.stringify(res.body));
-          setTimeout(() => {
-            this.spinner.hide();
-            this.modal.close();
-            this.router.navigate(["/room"]);
-          }, 2000);
-        },
-        err => {
-          setTimeout(() => {
-            this.spinner.hide();
-            localStorage.clear();
-            this.notifier.notify(
-              "error",
-              "Não foi possível cadastrar. Por favor, tente novamente"
-            );
-          }, 2000);
-        }
+    if (value.nome == "" || value.username == "" || value.email == "" || value.password == "") {
+      this.notifier.notify(
+        "error",
+        "Todos os dados são obrigatórios. Por favor, tente novamente"
       );
+    } else {
+      this.spinner.show();
+      this.user = new Usuario(
+        value.nome,
+        value.username,
+        value.email,
+        value.password,
+        0,
+        1,
+        'default.png',
+        true
+      );
+      this.httpClient
+        .post("http://monica:64803/api/Usuario", this.user, {
+          observe: "response"
+        })
+        .pipe(map(res => res as any))
+        .subscribe(
+          res => {
+            localStorage.setItem("userId", JSON.stringify(res.body));
+            setTimeout(() => {
+              this.spinner.hide();
+              this.modal.close();
+              this.router.navigate(["/room"]);
+            }, 2000);
+          },
+          err => {
+            setTimeout(() => {
+              this.spinner.hide();
+              localStorage.clear();
+              this.notifier.notify(
+                "error",
+                "Não foi possível cadastrar. Por favor, tente novamente"
+              );
+            }, 2000);
+          }
+        );
+    }
   }
 }
