+++ src/app/login-page/login-form/login-form.component.ts
@@ -37,32 +37,39 @@ export class LoginFormComponent {
 
   onSubmit(value: any) {
     event.preventDefault();
-    this.spinner.show();
-    this.requestLogin = new UsuarioLogin(value.username, value.senha);
-    this.httpClient
-      .post("http://monica:64803/api/Login", this.requestLogin, {
-        observe: "response"
-      })
-      .pipe(map(res => res as any))
-      .subscribe(
-        res => {
-          localStorage.setItem("userId", JSON.stringify(res.body));
-          setTimeout(() => {
-            this.spinner.hide();
-            this.modal.close();
-            this.router.navigate(["/room"]);
-          }, 2000);
-        },
-        err => {
-          setTimeout(() => {
-            this.spinner.hide();
-            localStorage.clear();
-            this.notifier.notify(
-              "error",
-              "Não foi possível cadastrar. Por favor, tente novamente"
-            );
-          }, 2000);
-        }
+    if (value.username == "" || value.senha == "") {
+      this.notifier.notify(
+        "error",
+        "Todos os dados são obrigatórios. Por favor, tente novamente"
       );
+    } else {
+      this.spinner.show();
+      this.requestLogin = new UsuarioLogin(value.username, value.senha);
+      this.httpClient
+        .post("http://monica:64803/api/Login", this.requestLogin, {
+          observe: "response"
+        })
+        .pipe(map(res => res as any))
+        .subscribe(
+          res => {
+            localStorage.setItem("userId", JSON.stringify(res.body));
+            setTimeout(() => {
+              this.spinner.hide();
+              this.modal.close();
+              this.router.navigate(["/room"]);
+            }, 2000);
+          },
+          err => {
+            setTimeout(() => {
+              this.spinner.hide();
+              localStorage.clear();
+              this.notifier.notify(
+                "error",
+                "Não foi possível cadastrar. Por favor, tente novamente"
+              );
+            }, 2000);
+          }
+        );
+    }
   }
 }
